Update and delete API Fixes

+++ User.Core/Services/UserCoreService.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using User.Core.DTOs;
 using User.Core.Interfaces.Core;
 using User.Core.Interfaces.Repo;
 using User.Repo;
@@ -28,8 +27,7 @@ namespace User.Core
         public async Task<IEnumerable<UserDTO>> GetUsers()
         {
             var users = await _userRepo.GetUsers().ConfigureAwait(false);
-            _mapper.Map<IEnumerable<UserDTO>>(users);
-            throw new BusinessException("This is business exception.");
+            return _mapper.Map<IEnumerable<UserDTO>>(users);
         }
 
         public async Task<UserDTO> CreateUser(UserDTO user)
@@ -41,12 +39,22 @@ namespace User.Core
         public async Task UpdateUser(int userId, UserDTO user)
         {
             var sysUser = await _userRepo.GetUser(userId).ConfigureAwait(false);
+
+            if (sysUser == null) throw new BusinessException("User not found.");
+
+            sysUser.FirstName = user.FirstName;
+            sysUser.Address = user.Address;
+            sysUser.Contact = user.Contact;
+
             await _userRepo.UpdateUser(sysUser).ConfigureAwait(false);
         }
 
         public async Task DeleteUser(int userId)
         {
             var sysUser = await _userRepo.GetUser(userId).ConfigureAwait(false);
+
+            if (sysUser == null) throw new BusinessException("User not found.");
+
             await _userRepo.DeleteUser(sysUser).ConfigureAwait(false);
         }
     }
